+++ lib/trailblazer/activity/schema.rb
@@ -11,5 +11,3 @@ module Trailblazer
     end # Schema
   end
 end
-require "trailblazer/activity/schema/implementation"
-require "trailblazer/activity/schema/intermediate"